Remove redundant guard layer by moving -—pins-guards to regrouping layer.
This fixes crashing POR + LTL.
The guards layer should not be used, as this functionality is better implemented by the language module. Remove?

The same holds probably for GBgetGroupPerm and GBgetColPerm. Since the language module gives the LTStype for the whole pins stack, there is no need to delegate any reordering to the regrouping layer.

Also removed insanely long if statement in regrouping layer.