Implement free for aligned memory on Windows.
Windows requires all memory that is allocated with _aligned_malloc,
to be free'd with _aligned_free. To this end RTalignedFree, and
HREalignedFree have been added.

So from now on, all memory allocated with RTalign and RTalignZero
has to be free'd with RTalignedFree, and all memory allocated
with HREalign and HREalignZero has to be free'd with
HREalignedFree.